add UNUSED(expr) makro to denote unused parameters
The issue with commenting out parameter names in function definitions is
that those cannot be properly be documented with Doxygen.
To omit compiler warnings about unsued parameters, one should write

(void)(param);

in the function body.
For ease of use and readability, a makro for this is defined similar to
the one defined by the Qt project (Q_UNUSED):

UNUSED(param);