cmake: rework 3rd-party library detection and reporting
By using the 'FeatureSummary' module we can easily produce a nice
summary of features and dependend libraries/packages at the end of each
CMake run.

However, to tell CMake that certain libraries/packages which have not
been found on the system will be downloaded and prepared automatically
on the first call to 'make', a dirty hack on CMakes (not documented)
internal handling of the 'find_package' command is required.
I hope somebody else or the future me can actually understand what I was
doing right now ...