examples: boris: linearized particle cloud components
Previously, the positions and velocities of a particle cloud were a
2D nested array with particle index as first and axis index as second
index.
This would cause some significant overhead and complexity when it comes
to parallelization and communication.

Now, the positions and velocities are linear vectors with the
coordinates of the particles:

    x1, y1, z1, x2, y2, z2, x3, y3, z3, ..., xn, yn, zn
    u1, v1, w1, u2, v2, w2, u3, v3, w3, ..., un, vn, wn

While I've been at it, some for loops have been replaced by
iterator-based STL functions.