config: enabled cmd-based log-file specification
Introducing a new configuration option `log_prefix` to specify a (base)
name or prefix for the log files.
With MPI enabled, this is a prefix and without MPI this will become the
complete log file name (appended by the file extension `.log`).

Also refactored a little bit how the loggers are set up.

Signed-off-by: Torbjörn Klatt <t.klatt@fz-juelich.de>