start scaffold for v2
The quadrature stuff is mostly coppied from the old PFASST++, while
Encapsulation, Sweeper, Controller and Transfer are newly designed with
the help of templates and type traits for static polymorphism instead of
runtime polymorphism.

As the sweeper is now Q-matrix-centric and does not distinguesh between
the initial value and the first node, the Q-matrices are now of the size
M+1xM+1.
Note: For now, only Gauss-Radau-right has been adjusted for this. The
other quadrature types are not tested!